[BEAM-10135][BEAM-10136] Refactor jdbc external transform registrar
